feat: enhance PlaybackControls with SeekStepControls and update seek step validation
