+++ packages/studio-base/src/components/PlaybackControls/sharedHelpers.ts
@@ -14,7 +14,7 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import { Time, toMillis, fromMillis } from "@foxglove/rostime";
+import { Time, toSec, fromSec } from "@foxglove/rostime";
 
 const ARROW_SEEK_BIG_MS = 500;
 const ARROW_SEEK_DEFAULT_MS = 100;
@@ -30,7 +30,7 @@ export const jumpSeek = (
   modifierKeys?: { altKey: boolean; shiftKey: boolean },
   defaultMsOverride?: number,
 ): Time => {
-  const timeMs = toMillis(currentTime);
+  const timeSec = toSec(currentTime);
   const defaultMs =
     defaultMsOverride != undefined && defaultMsOverride > 0
       ? defaultMsOverride
@@ -41,5 +41,5 @@ export const jumpSeek = (
       : modifierKeys?.shiftKey === true
       ? ARROW_SEEK_SMALL_MS
       : defaultMs;
-  return fromMillis(timeMs + deltaMs * directionSign);
+  return fromSec(timeSec + (deltaMs / 1000) * directionSign);
 };
