+++ packages/studio-base/src/i18n/cosEn/general.ts
@@ -24,6 +24,6 @@ export const cosGeneral = {
   seekForward: "Seek forward",
   seekStep: "Seek step",
   loopPlayback: "Loop playback",
-  invalidSeekStep: "Seek step must be > 1e-9s and < 1 hour. Reset to 0.1 s.",
+  invalidSeekStep: "Seek step must be >= 1e-6s and <= 1 hour. Reset to 0.1 s.",
   playbackSpeed: "Playback speed",
 };
